functions/raw-fetch-yfinance/requirements.txt
functions-framework==3.*
google-cloud-bigquery
yfinance
pandas
google-cloud-secret-manager
duckdb==1.3.2
pyarrow
requests
google-cloud-storage
pyyaml
tqdm